move tests to __tests__ directory
