+++ packages/flex-plugin/src/utils/__tests__/runtime.test.ts
@@ -1,4 +1,4 @@
-import { getAssetsUrl } from './runtime';
+import { getAssetsUrl } from '../runtime';
 
 describe('runtime', () => {
     describe('getAssetsUrl', () => {
